backend/requirements.txt
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
fastapi==0.109.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.1.3
langchain-core==1.2.0
langchain-ollama==0.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.0
langsmith==0.4.59
numpy==1.24.3
openai-whisper==20231117
orjson==3.11.5
ormsgpack==1.12.0
packaging==25.0
Pillow==10.2.0
pydantic==2.12.5
pydantic-settings==2.1.0
pydantic_core==2.41.5
PyMuPDF==1.23.8
python-docx==1.1.0
python-pptx==0.6.23
pytesseract==0.3.10
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.27.0
xxhash==3.6.0
yt-dlp==2023.12.30
zstandard==0.25.0
